Move esp32 specific gen to esp32_helper
